requirements.txt
accelerate==0.34.2
aenum==3.1.15
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
awscli==1.32.80
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.80
botocore==1.34.80
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
contourpy==1.3.0
ctransformers==0.2.27
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.21.0
deepspeed==0.15.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docker-pycreds==0.4.0
docutils==0.16
einops==0.8.0
evaluate==0.4.2
exceptiongroup==1.2.2
faiss-cpu==1.8.0
filelock==3.16.0
Flask==3.0.3
Flask-Cors==4.0.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
gradientai==1.13.1
greenlet==3.1.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.52
langchain-openai==0.1.1
langchain-text-splitters==0.0.2
langsmith==0.1.117
lightning-utilities==0.11.7
llama-index==0.10.27
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.68.post1
llama-index-embeddings-gradient==0.1.2
llama-index-embeddings-langchain==0.1.2
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48.post3
llama-index-llms-anyscale==0.1.4
llama-index-llms-langchain==0.1.3
llama-index-llms-openai==0.1.31
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
llamaindex-py-client==0.1.19
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.9.1
nodelist-inflator==0.2.0
nomic==3.1.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openai==1.44.1
orjson==3.10.7
packaging==23.2
pandas==2.2.2
peft==0.12.0
pillow==10.4.0
platformdirs==4.3.2
protobuf==5.28.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.4
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.8.1
rsa==4.7.2
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==2.6.1
sentencepiece==0.2.0
sentry-sdk==2.14.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
striprtf==0.0.26
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.1
torchmetrics==1.4.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.9
Werkzeug==3.0.4
wrapt==1.16.0
xxhash==3.5.0
yarl==1.11.1
